curative effects of milk bathing
