+++ update.json
@@ -188,6 +188,6 @@
 	,"1.5.11":"Stop cyclic machines trying to work while chunks are unloaded, causing TPS issues on servers.  Default config values changed (affects new and not existing config files): For both end and nether ores, spawnChanceDiamond = 1 and spawnChanceEmerald = 1.   For Uncrafter: added elytra and techreborn:uumatter to both BlacklistOutput and BlacklistInput.  Added resourcefulcrops to BlacklistMod.  Removed 'Elytra Repair' config entry and feature (overpowered and not working right).  Monster ball is now reuseable; does not break after a single use.   Fixed Ender Wings and Ender Book not playing sounds on teleport.  Made small changes to about five recipes.   Added support for Enchantment Descriptions (which are added by walia and other mods).  Fix BeheadingExtraMobs config system for the enchantment, it now uses modname:monster-player, matching the summon command.  Fix buttons on some GUI's having broken textures on the bottom.  Automated User in attack mode will no longer hit entities outside of its bounding box. "
 	,"1.6.0":"No longer uses JSON recipe files, this fixes the error in forge that removes ALL mod recipes when any one item is disabled in config.  All machines in cyclic that have fuel slots will no longer run for free, they now require fuel.  Fixed some tooltip issues.  The Ender Book and Ender Wing items now hook into forge event system, so they can get affected/blocked/redirected other items/blocks (compatibility with DarkUtilities and other mods).  Interfaces with textboxes now allow you to hold down the button (Ender Book, Faith Plate, Password Detector).  Automated User can now attack tiny entities such as silverfish.  Reduced cooldown on Exchange scepters.  Fixed wither skeleton skulls not dropping from Beheading enchantment.  Beheading enchant will no longer crash the game when used on a player.  All new custom made sounds for Ender Wings and Ender Book.  Fixed a potential exploit in Unchant Pylon.   Experience pylon range increased to 8 blocks away, up from 5.  Fixed Experience pylon collection area; it is now centered on the block instead of only working on one side.   "
 	,"1.6.1":"You can now sneak to cancel sleeping in the Sleeping Mat.  Fix item tooltips not showing in Sack of Storage.  Fix Entity Detector range being offcenter; also added a preview button to render selected area.  Auto Fisher should no longer let fish and other loot into fishing rod slot when the rod breaks.  Fixed some missing recipes.  Charms can now be repaired in an anvil instead of crafting table.  The horse upgrade carrots will now get eaten only one at a time, instead of double. "
-	,"1.6.2":"Controlled miner now accepts items into the tool slot from above (hoppers & other automation - thanks to CatAndHorse).  New block: Hydrator for making concrete, and other blocks.  Doubled the damage output of the Evoker Fang.  Purple ability armor no longer takes chain armor to craft, and has similar defense to diamond.  New Item: Apple of Lofty Stature.  New Item: Inverted Stirrups.  Bucket Storage block has been overhauled and replaced by a new block: Fluid Storage Tank that takes any fluid.   Registered Milk and Experience in forge Fluid Registry, which means they can be put into fluid tanks from a bucket.  Experience Pylon overhauled to now use experience fluid (WARNING: existing placed pylons during upgrade may lose xp).  "
+	,"1.6.2":"Controlled miner now accepts items into the tool slot from above (hoppers & other automation - thanks to CatAndHorse).  New block: Hydrator for making concrete, and other blocks.  Doubled the damage output of the Evoker Fang.  Purple ability armor no longer takes chain armor to craft, and has similar defense to diamond.  New Item: Apple of Lofty Stature.  New Item: Inverted Stirrups.  Bucket Storage block has been overhauled and replaced by a new block: Fluid Storage Tank that takes any fluid.   Registered Milk and Experience in forge Fluid Registry, which means they can be put into fluid tanks from a bucket.  Bathing in milk will cure potion effects just like drinking it.  Experience Pylon overhauled to now use experience fluid (WARNING: existing placed pylons during upgrade may lose xp).  "
   }
 }
